add controllers, change services and repositories

+++ src/main/java/jr_course/service/GrammarServiceImpl.java
@@ -2,10 +2,12 @@ package jr_course.service;
 
 import jr_course.dao.GrammarRepository;
 import jr_course.entity.Grammar;
+import jr_course.entity.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import jr_course.exception.*;
 
 import java.util.List;
 import java.util.Optional;
@@ -29,9 +31,12 @@ public class GrammarServiceImpl implements GrammarService {
     }
 
     @Override
-    public List<Grammar> findByLevel(int level) {
+    public List<Grammar> findAllByLevel(int level) {
         logger.info("\"findAllByLevelContains(level)\"");
         logger.info("Find all words in DB depending on the \"" + level + "\" level.");
+        if (!(level >= 1 && level <= 5))
+            throw new IncorrectDataInputException("Incorrect level data input - " + level);
+
         return grammarRepository.findAllByLevel(level);
     }
 
@@ -46,8 +51,8 @@ public class GrammarServiceImpl implements GrammarService {
         if (result.isPresent()) {
             grammar = result.get();
         } else {
-            logger.warn("Grammar not found.");
-            throw new RuntimeException("Грамматика с id - " + id + " не была найдена.");
+            logger.warn("Grammar was not found.");
+            throw new DataNotFoundException("Grammar with id " + id + " was not found.");
         }
 
         logger.info("Return grammar.");
@@ -64,8 +69,66 @@ public class GrammarServiceImpl implements GrammarService {
     @Override
     public void deleteById(int id) {
         logger.info("\"deleteById(id)\"");
-        logger.info("Delete a grammar in DB by id " + id + ".");
-        grammarRepository.deleteById(id);
+        logger.info("Delete a grammar by id " + id + ".");
+        if (grammarRepository.existsById(id)) grammarRepository.deleteById(id);
+        else throw new DataNotFoundException("Grammar with id \" + id + \" was not found.");
     }
 
+    @Override
+    public List<Grammar> findAllByUserId(int userId) {
+        logger.info("\"findAllByUserCollection_Id(userId)\"");
+        logger.info("Find all grammar by user id " + userId + ".");
+        return grammarRepository.findAllByUserCollection_Id(userId);
+    }
+
+    @Override
+    public void addGrammarToPersonalList(User user, int grammarId) {
+        logger.info("\"addGrammarToPersonalList(user, grammarId)\"");
+        logger.info("Add grammar to personal grammar list.");
+
+        Grammar grammar = findById(grammarId);
+
+        if (grammar.getUserCollection().contains(user)) {
+            logger.info("Grammar was already added!");
+        } else {
+            logger.info("Add grammar.");
+            grammar.addUser(user);
+
+            logger.info("Update word.");
+            grammarRepository.save(grammar);
+            logger.info("Grammar was added to personal grammar list.");
+        }
+    }
+
+    @Override
+    public void deleteGrammarFromPersonalList(int grammarId, User user) {
+        logger.info("\"deleteGrammarFromPersonalList(grammarId, user)\"");
+        logger.info("Delete grammar by id " + grammarId + "from personal grammar list.");
+
+        logger.info("Find grammar by id and delete user.");
+        Grammar grammar = findById(grammarId);
+
+        if (grammar.getUserCollection().contains(user)) {
+            grammar.deleteUser(user);
+
+            logger.info("Update grammar.");
+            grammarRepository.save(grammar);
+            logger.info("Grammar was deleted from personal grammar list!");
+        } else logger.info("Grammar with id + " + grammarId + " was not found in the personal list." );
+    }
+
+    @Override
+    public List<Grammar> findByDifferentParameters(String param) {
+        logger.info("\"findByDifferentParameters(param)\"");
+        logger.info("Find grammar by parameter " + param + ".");
+
+        param = param.trim();
+        List<Grammar> grammarList = grammarRepository
+                .findByFormulaContainsOrExampleContainsAllIgnoreCase(param, param);
+
+        if (grammarList.isEmpty()) logger.info("Grammar list is empty.");
+
+        logger.info("Return grammar list.");
+        return grammarList;
+    }
 }
