+++ src/main/java/jr_course/service/NoteServiceImpl.java
@@ -1,11 +1,13 @@
 package jr_course.service;
 
 import jr_course.dao.NoteRepository;
+import jr_course.dao.UserRepository;
 import jr_course.entity.Note;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import jr_course.exception.*;
 
 import java.util.List;
 
@@ -13,16 +15,20 @@ import java.util.List;
 public class NoteServiceImpl implements NoteService {
     private Logger logger = LoggerFactory.getLogger(this.getClass().getName());
     private NoteRepository noteRepository;
+    private UserRepository userRepository;
 
     @Autowired
-    public NoteServiceImpl(NoteRepository noteRepository) {
+    public NoteServiceImpl(NoteRepository noteRepository, UserRepository userRepositor) {
         this.noteRepository = noteRepository;
+        this.userRepository = userRepositor;
     }
 
     @Override
-    public List<Note> findAllByUser_Id(int id) {
-        logger.info("\"findAllByUser_Id(id)\"");
+    public List<Note> findAllByUserId(int id) {
+        logger.info("\"findAllByUserId(id)\"");
         logger.info("Find all words by user id " + id + ".");
+        if (!userRepository.existsById(id))
+            throw new DataNotFoundException("User with id " + id + " was not found.");
         return noteRepository.findAllByUser_Id(id);
     }
 
@@ -37,6 +43,9 @@ public class NoteServiceImpl implements NoteService {
     public void deleteById(int id) {
         logger.info("\"deleteById(id)\"");
         logger.info("Delete a note by id " + id + ".");
-        noteRepository.deleteById(id);
+        if (noteRepository.existsById(id)) {
+            noteRepository.deleteById(id);
+            logger.info("Note was deleted!");
+        } else throw new DataNotFoundException("Note with id " + id + " was not found.");
     }
 }
